backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.18
pydantic==2.10.6
python-dotenv==1.1.0